libs/common/requirements.txt
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
structlog==24.4.0
google-api-python-client==2.136.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
python-slugify==8.0.4
tzdata==2024.1
babel==2.15.0


